chore: moves storage services to services folder
